Apply changes from main to gh-pages via copying strategy
This applies the same copying files strategy used in the preview
builds to main, so that the unrelated histories of the gh-pages
and main branches don't need to be reconciled.

It also adds extra info to each commit applied to gh-pages in this
way so that they can be related back to what was merged into main,
if needed.