Use flexbox to size and align items in the prefooter
This simplifies a bunch of styles that were using margins to
just use flex and gap, so that we don't have margins on individual
list items or differing top/bottom margins.

Also adjusts the breakpoints at which the footer changes layout
in order to more closely match the designs (as well as the SUL
footer below it).